pyproject.toml
[project]
name = "livekit-python-agent"
version = "1.0.0"
description = "LiveKit Python Agent with LiveAvatar integration, inspired by playground agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero]~=1.3.6",
    "livekit-plugins-google>=1.3.6",
    "livekit-plugins-liveavatar>=1.3.6",
    "python-dotenv>=1.2.1",
]

[project.scripts]
agent = "src.agent:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
